Add files via upload
The agent is done. 
The game is done. 
The saver contains the save function. Make some changes on it. 
The other two functions in the helpe has not yet been used.
The loss in train is incorrect.
The save in train is not set.
The play file has not been run.
The play file needs a method to give the output.